feat: enhance SCPF action with improved error handling and optional patterns
- Update workflow to continue on PR creation errors and add failure check
- Simplify scan command in action.yml and make SARIF upload continue on errors
- Remove input type constraints in action.yml for better flexibility
- Mark upgradeable contract patterns as optional with skip on error to avoid false positives
- Add validation to missing-initializer-modifier pattern for robustness

This improves reliability of the CI pipeline and reduces unnecessary failures while maintaining security checks.

+++ crates/scpf-core/src/semantic.rs
@@ -55,9 +55,9 @@ impl SemanticScanner {
                 );
 
                 let lines: Vec<&str> = pattern.pattern.lines().collect();
-                if (e.row as usize) < lines.len() {
-                    tracing::error!("Error line: {}", lines[e.row as usize]);
-                    tracing::error!("Position:   {}^", " ".repeat(e.column as usize));
+                if e.row < lines.len() {
+                    tracing::error!("Error line: {}", lines[e.row]);
+                    tracing::error!("Position:   {}^", " ".repeat(e.column));
                 }
 
                 return Err(anyhow::anyhow!(
